feat: resolved problem with code duplication
